New: Extension for splitting addresses into chunks.
